Remove fresh assertCountStatementIterator (reason below)
The function assertCountStatementIterator basically
counts how many elements are in an iterator-instance.
It later compares that number with an expected one and
checks them using assertEquals.

I simplified that. SetResult has now a count function.
Furthermore, all assertCountStatementIterator checks
were replaced with a simple $statements->count() call
and an assertEquals. Now its more clear for the user
what the test does at this point, i hope :)

If it is neccessary in the future to have a special
count function for huge datasets (using paging etc.),
we can implement that in the according iterator-class.