Change createGraph/dropGraph parameter in Store.php
Given $graph must be of type NamedNode, not just Node.
Reason is, that we want a graph which is represented
by an URI, so we dont want Literals or AnyPattern given.

Drop createGraph and dropGraph from all store implementations,
because they basically call the query method. These 2
methods are now located in AbstractSparqlStore.php.