Remove ChainFactory and StoreFactory #17
StoreFactory forces the use of a DI-container, because
our Store implementations have a different set of
dependencies for their constructor. (QueryCache vs. Virtuoso)

Because we dont want to introduce Dice at this level,
we are forced to let the ChainFactory and StoreFactory die
here and ressurrect it at application level, if needed.

ChainFactory was removed because it relies on the
StoreFactory.