Introduce ResultFactory; Port Result classes to Impl-way #17 #34
ResultFactory from now on will be used to create SPARQL store
results. (#17)

Port Result classes to Impl-way. All classes of Saft\Store\Result
were not using the naming with Impl as suffix for concrete
implementations (class). E.g. we used SetResult directly as an
instance, but from now on SetResult is an interface and we only
get an instance of it by using ResultFactoryImpl's methods. (#34)

Port all classes and tests.