StoreAbstractTest: correct an existing test and add a new one
regarding to testGetMatchingStatementsCheckForTriples:

it iterates about ALL triples of the whole store because graph was
undefined, so it used the default graph. if the store contains 10k
triples, it would iterated over all of them .... :)

changed that to check only one triple, if default graph and add added a
new test to pull from a certain graph only one quad and check that later
on.